[APPC-1248] Moved referral files to specific package
