+++ app/src/main/java/com/asfoundation/wallet/di/ToolsModule.java
@@ -97,7 +97,9 @@ import com.asfoundation.wallet.poa.ProofWriter;
 import com.asfoundation.wallet.poa.TaggedCompositeDisposable;
 import com.asfoundation.wallet.promotions.PromotionsInteractorContract;
 import com.asfoundation.wallet.promotions.PromotionsTestInteractor;
-import com.asfoundation.wallet.promotions.SharedPreferencesReferralLocalData;
+import com.asfoundation.wallet.referrals.ReferralInteractorContract;
+import com.asfoundation.wallet.referrals.ReferralTestInteractor;
+import com.asfoundation.wallet.referrals.SharedPreferencesReferralLocalData;
 import com.asfoundation.wallet.repository.ApproveService;
 import com.asfoundation.wallet.repository.ApproveTransactionValidatorBds;
 import com.asfoundation.wallet.repository.BalanceService;
